app/requirements-windows.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
defusedxml==0.7.1
distro==1.9.0
Django==5.1.6
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
h11==0.14.0
httpcore==1.0.7
httpie==3.2.4
httpx==0.28.1
idna==3.10
jiter==0.8.2
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.1.0
openai==1.65.2
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.9.0
PySocks==1.7.1
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
